refactor: restructure project into monorepo
