+++ apps/web/package.json
@@ -1,0 +1,68 @@
+{
+  "name": "care-ai",
+  "version": "1.0.0",
+  "description": "Assistente AI pessoal para gerenciar rotina diária",
+  "private": true,
+  "scripts": {
+    "dev": "next dev",
+    "build": "prisma generate --schema ../../backend/prisma/schema.prisma && next build",
+    "start": "next start",
+    "lint": "next lint",
+    "type-check": "tsc --noEmit",
+    "db:generate": "prisma generate --schema ../../backend/prisma/schema.prisma",
+    "db:push": "prisma db push --schema ../../backend/prisma/schema.prisma",
+    "db:migrate": "prisma migrate dev --schema ../../backend/prisma/schema.prisma",
+    "db:studio": "prisma studio --schema ../../backend/prisma/schema.prisma",
+    "db:seed": "tsx ../../backend/prisma/seed.ts",
+    "postinstall": "prisma generate --schema ../../backend/prisma/schema.prisma"
+  },
+  "dependencies": {
+    "@headlessui/react": "^1.7.0",
+    "@heroicons/react": "^2.0.0",
+    "@hookform/resolvers": "^3.3.0",
+    "@next-auth/prisma-adapter": "^1.0.7",
+    "@prisma/client": "^5.6.0",
+    "@tailwindcss/forms": "^0.5.0",
+    "@types/bcryptjs": "^2.4.6",
+    "@types/jsonwebtoken": "^9.0.0",
+    "@types/node": "^20.8.0",
+    "@types/nodemailer": "^6.4.0",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
+    "@types/uuid": "^9.0.0",
+    "autoprefixer": "^10.4.0",
+    "bcryptjs": "^2.4.3",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.0.0",
+    "date-fns": "^2.30.0",
+    "framer-motion": "^10.16.0",
+    "groq-sdk": "^0.23.0",
+    "ioredis": "^5.3.0",
+    "jsonwebtoken": "^9.0.0",
+    "lucide-react": "^0.288.0",
+    "next": "^14.0.0",
+    "next-auth": "^4.24.0",
+    "nodemailer": "^6.9.0",
+    "openai": "^4.20.0",
+    "postcss": "^8.4.0",
+    "prisma": "^5.6.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.47.0",
+    "react-hot-toast": "^2.4.0",
+    "react-icons": "^5.5.0",
+    "recharts": "^2.8.0",
+    "swr": "^2.2.0",
+    "tailwindcss": "^3.3.0",
+    "typescript": "^5.2.0",
+    "uuid": "^9.0.0",
+    "zod": "^3.22.0"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^6.9.0",
+    "@typescript-eslint/parser": "^6.9.0",
+    "eslint": "^8.52.0",
+    "eslint-config-next": "^14.0.0",
+    "tsx": "^4.0.0"
+  }
+}
