+++ apps/web/public/manifest.json
