+++ apps/web/src/app/api/chat/route.ts
