+++ apps/web/src/assets/images/README.md
