+++ apps/web/src/components/FreechAvatar.tsx
