+++ apps/web/src/components/Tasks.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useState, useEffect } from 'react'
-import { Task } from '@/types'
+import { Task } from '@shared/types'
 import { formatDate, getPriorityColor } from '@/lib/utils'
 import {
   PlusIcon,
