+++ apps/web/src/lib/agent-functions.ts
