+++ apps/web/src/lib/groq.ts
