+++ apps/web/src/lib/openai.ts
