+++ shared/utils.ts
@@ -1,0 +1,3 @@
+export function generateId(): string {
+  return Math.random().toString(36).substr(2, 9);
+}
